fix(reborn): make Slack + OpenAI-compat host-beta build and wire after WebUI merge
The WebUI host-stack merge (parent commit) hoisted `webui_v2_app` + its config,
authenticator, and middleware surface from `ironclaw_reborn_composition` up into
`ironclaw_reborn_webui_ingress`, but left the `slack-v2-host-beta` /
`openai-compat-beta` blocks in the moved `webui_serve.rs` pointing at
`crate::slack::*` / composition internals that don't exist in ingress. Those
features were declared only as known-cfgs and compiled out, so:

- composition failed to build under `slack-v2-host-beta` (two mount-vocabulary
  imports still on the old `webui::webui_serve` path);
- the ingress serve blocks were permanently dead, so the CLI's slack/openai
  features forwarded to composition but never mounted the Slack routes —
  a functional parity break, not just a compile break;
- composition's slack-gated tests still imported the moved `webui_v2_app`.

Wiring (behavior-preserving; restores pre-merge parity):
- ingress now defines real `slack-v2-host-beta` / `openai-compat-beta` features
  that forward to composition (+ optional `ironclaw_reborn_openai_compat`); the
  moved `webui_serve.rs` reaches Slack setup/route types and the
  OpenAI-compat bearer-evidence helper through composition's public facade
  (`ironclaw_reborn_composition::{SlackPersonalSetupServiceSlot,
  SlackChannelRouteAdminRouteConfig, slack_channel_route_admin_route_mount,
  SlackPersonalOAuthBindingConfig, mark_bearer_token_verified_for_tenant}`).
  Ingress does NOT depend on `ironclaw_product_adapters` directly — the
  architecture boundary (`reborn_dependency_boundaries.rs`) forbids it, so the
  evidence helper is re-exported from composition instead.
- composition promotes `slack_channel_route_admin_route_mount` + its
  `SlackChannelRouteAdminRouteMount` return type to `pub` (its sole caller,
  `webui_v2_app`, moved up), mirroring the already-public `ProtectedRouteMount`.
- CLI forwards `slack-v2-host-beta` / `openai-compat-beta` to the ingress crate
  as well as composition, so the serve blocks compile in and the routes mount.

Tests:
- The 7 composition slack unit tests that drove the now-relocated `webui_v2_app`
  move to `ironclaw_reborn_webui_ingress/tests/slack_host_beta_webui_v2.rs`.
  They use only composition's public builders, so ingress (which normal-deps
  composition — single crate copy, no dev-dep cycle) is their correct home; a
  composition lib-test cannot call the ingress `webui_v2_app` without cargo
  building two incompatible copies of composition. composition's ingress
  dev-dep gains `slack-v2-host-beta` so its own `webui_v2_product_auth*` tests'
  `with_slack_*` blocks compile.

Verified (clean env): composition/ingress/cli build + clippy `-D warnings` under
both beta features; ingress `--all-features` suite green incl. the 7 relocated
tests; composition lib (1589) + router (webui_v2_serve 44 / product_auth 51) +
cli (440 incl. Dockerfile smoke) green; `ironclaw_architecture` boundaries hold.
