+++ crates/ironclaw_reborn_composition/src/slack/slack_host_beta.rs
@@ -80,7 +80,7 @@ use crate::slack::slack_serve::{
     SlackInstallationSelector, SlackTeamId, SlackUserId, StaticSlackInstallationResolver,
     slack_events_route_mount,
 };
-use crate::webui::webui_serve::PublicRouteMount;
+use crate::webui::route_mounts::PublicRouteMount;
 
 const SLACK_BOT_TOKEN_HANDLE: &str = "slack_bot_token";
 const SLACK_SIGNATURE_HEADER: &str = "X-Slack-Signature";
@@ -1359,10 +1359,9 @@ mod tests {
     use ironclaw_processes::{InMemoryProcessResultStore, InMemoryProcessStore, ProcessServices};
     use ironclaw_product_workflow::{
         LifecyclePackageKind, LifecyclePackageRef, ProductActorUserResolutionRequest,
-        ProductWorkflowError, RebornChannelConnectStrategy, RebornExtensionOnboardingState,
-        RebornOutboundDeliveryTargetId, RebornOutboundDeliveryTargetStatus,
-        RebornServicesErrorCode, RebornServicesErrorKind, RebornSetOutboundPreferencesRequest,
-        WebUiAuthenticatedCaller,
+        ProductWorkflowError, RebornExtensionOnboardingState, RebornOutboundDeliveryTargetId,
+        RebornOutboundDeliveryTargetStatus, RebornServicesErrorCode, RebornServicesErrorKind,
+        RebornSetOutboundPreferencesRequest, WebUiAuthenticatedCaller,
     };
     use ironclaw_resources::InMemoryResourceGovernor;
     use ironclaw_secrets::InMemorySecretStore;
@@ -1380,8 +1379,7 @@ mod tests {
     use crate::slack::slack_channel_routes::{
         InMemorySlackChannelRouteStore, SlackChannelRoute, SlackChannelRouteAdminRouteMount,
         SlackChannelRouteError, SlackChannelRouteKey, SlackChannelRouteListPage,
-        WEBUI_V2_CHANNELS_SLACK_ALLOWED_PATH, WEBUI_V2_CHANNELS_SLACK_ROUTES_PATH,
-        slack_channel_route_admin_route_mount,
+        WEBUI_V2_CHANNELS_SLACK_ROUTES_PATH, slack_channel_route_admin_route_mount,
     };
     use crate::slack::slack_connectable_channel::{
         SlackOperatorRouteVisibility, build_webui_services_with_slack_host_beta_mounts,
@@ -1395,8 +1393,7 @@ mod tests {
     use crate::slack::slack_serve::{SlackApiAppId, SlackUserId};
     use crate::{
         RebornBuildError, RebornBuildInput, RebornRuntimeIdentity, RebornRuntimeInput,
-        SLACK_EVENTS_PATH, WebuiAuthentication, WebuiAuthenticator, WebuiServeConfig,
-        build_reborn_runtime, local_dev_runtime_policy, webui_v2_app,
+        SLACK_EVENTS_PATH, build_reborn_runtime, local_dev_runtime_policy,
     };
 
     const TENANT: &str = "tenant:slack-host";
@@ -1445,61 +1442,6 @@ mod tests {
     #[should_panic(expected = "trigger run failed before acceptance")]
     fn accepted_trigger_run_id_rejects_pre_acceptance_error() {
         accepted_trigger_run_id(None, Some(TriggerRunHistoryStatus::Error));
-    }
-
-    struct OperatorTokenAuthenticator;
-
-    #[async_trait]
-    impl WebuiAuthenticator for OperatorTokenAuthenticator {
-        async fn authenticate(&self, token: &str) -> Option<WebuiAuthentication> {
-            if token == "operator-token" {
-                Some(WebuiAuthentication::operator(
-                    UserId::new(USER).expect("user"),
-                ))
-            } else {
-                None
-            }
-        }
-
-        fn mounts_operator_webui_config_routes(&self) -> bool {
-            true
-        }
-    }
-
-    struct MixedSessionAndOperatorAuthenticator;
-
-    #[async_trait]
-    impl WebuiAuthenticator for MixedSessionAndOperatorAuthenticator {
-        async fn authenticate(&self, token: &str) -> Option<WebuiAuthentication> {
-            match token {
-                "session-token" => {
-                    Some(WebuiAuthentication::user(UserId::new(USER).expect("user")))
-                }
-                "operator-token" => Some(WebuiAuthentication::operator(
-                    UserId::new(USER).expect("user"),
-                )),
-                _ => None,
-            }
-        }
-
-        fn mounts_operator_webui_config_routes(&self) -> bool {
-            true
-        }
-    }
-
-    struct HiddenOperatorRouteAuthenticator;
-
-    #[async_trait]
-    impl WebuiAuthenticator for HiddenOperatorRouteAuthenticator {
-        async fn authenticate(&self, token: &str) -> Option<WebuiAuthentication> {
-            if token == "operator-token" {
-                Some(WebuiAuthentication::operator(
-                    UserId::new(USER).expect("user"),
-                ))
-            } else {
-                None
-            }
-        }
     }
 
     #[derive(Debug)]
@@ -2107,446 +2049,6 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn slack_allowed_channels_are_reachable_through_webui_v2_app() {
-        let (runtime, _root) = runtime().await;
-        let mounts = build_slack_host_beta_mounts(&runtime, config_without_channel_routes())
-            .expect("mounts");
-        let bundle = build_webui_services_with_slack_host_beta_mounts(
-            &runtime,
-            None,
-            Some(&mounts),
-            SlackOperatorRouteVisibility::Visible,
-        )
-        .expect("webui bundle");
-        let app = webui_v2_app(
-            bundle,
-            WebuiServeConfig::new(
-                TenantId::new(TENANT).expect("tenant"),
-                Arc::new(OperatorTokenAuthenticator),
-                Vec::new(),
-            )
-            .with_default_agent_id(AgentId::new(AGENT).expect("agent"))
-            .with_default_project_id(ProjectId::new(PROJECT).expect("project"))
-            .with_slack_channel_routes(mounts.channel_routes),
-        )
-        .expect("webui app");
-
-        let save = app
-            .clone()
-            .oneshot(
-                Request::builder()
-                    .method("PUT")
-                    .uri(WEBUI_V2_CHANNELS_SLACK_ALLOWED_PATH)
-                    .header("authorization", "Bearer operator-token")
-                    .header("content-type", "application/json")
-                    .body(Body::from(r#"{"channel_ids":["C0HOST","C0OPS"]}"#))
-                    .expect("save request builds"),
-            )
-            .await
-            .expect("save route responds");
-        assert_eq!(save.status(), StatusCode::OK);
-        let save_body = axum::body::to_bytes(save.into_body(), 64 * 1024)
-            .await
-            .expect("save body");
-        let save_body: serde_json::Value = serde_json::from_slice(&save_body).expect("save json");
-        assert_eq!(save_body["channels"].as_array().expect("channels").len(), 2);
-        assert_ne!(
-            save_body["channels"][0]["subject_user_id"],
-            save_body["channels"][1]["subject_user_id"],
-            "allowed API should assign one tenant-scoped subject per channel"
-        );
-
-        let list = app
-            .oneshot(
-                Request::builder()
-                    .method("GET")
-                    .uri(WEBUI_V2_CHANNELS_SLACK_ALLOWED_PATH)
-                    .header("authorization", "Bearer operator-token")
-                    .body(Body::empty())
-                    .expect("list request builds"),
-            )
-            .await
-            .expect("list route responds");
-        assert_eq!(list.status(), StatusCode::OK);
-        let body = axum::body::to_bytes(list.into_body(), 64 * 1024)
-            .await
-            .expect("list body");
-        let body: serde_json::Value = serde_json::from_slice(&body).expect("list json");
-        assert_eq!(
-            body["channels"],
-            serde_json::json!([
-                {
-                    "channel_id":"C0HOST",
-                    "subject_user_id": save_body["channels"][0]["subject_user_id"].clone(),
-                    "subject_display_name": save_body["channels"][0]["subject_display_name"].clone()
-                },
-                {
-                    "channel_id":"C0OPS",
-                    "subject_user_id": save_body["channels"][1]["subject_user_id"].clone(),
-                    "subject_display_name": save_body["channels"][1]["subject_display_name"].clone()
-                }
-            ])
-        );
-
-        runtime.shutdown().await.expect("runtime shuts down");
-    }
-
-    #[tokio::test]
-    async fn slack_connectable_channels_advertise_admin_action_to_operator_token() {
-        let (runtime, _root) = runtime().await;
-        let mounts = build_slack_host_beta_mounts(&runtime, config_without_channel_routes())
-            .expect("mounts");
-        let bundle = build_webui_services_with_slack_host_beta_mounts(
-            &runtime,
-            None,
-            Some(&mounts),
-            SlackOperatorRouteVisibility::Visible,
-        )
-        .expect("webui bundle");
-        let app = webui_v2_app(
-            bundle,
-            WebuiServeConfig::new(
-                TenantId::new(TENANT).expect("tenant"),
-                Arc::new(OperatorTokenAuthenticator),
-                Vec::new(),
-            )
-            .with_default_agent_id(AgentId::new(AGENT).expect("agent"))
-            .with_default_project_id(ProjectId::new(PROJECT).expect("project"))
-            .with_slack_channel_routes(mounts.channel_routes),
-        )
-        .expect("webui app");
-
-        let response = app
-            .oneshot(
-                Request::builder()
-                    .method("GET")
-                    .uri("/api/webchat/v2/channels/connectable")
-                    .header("authorization", "Bearer operator-token")
-                    .body(Body::empty())
-                    .expect("request builds"),
-            )
-            .await
-            .expect("route responds");
-
-        assert_eq!(response.status(), StatusCode::OK);
-        let body = axum::body::to_bytes(response.into_body(), 64 * 1024)
-            .await
-            .expect("response body");
-        let body: serde_json::Value = serde_json::from_slice(&body).expect("response json");
-        let strategies: Vec<_> = body["channels"]
-            .as_array()
-            .expect("channels")
-            .iter()
-            .map(|channel| channel["strategy"].as_str().expect("strategy"))
-            .collect();
-        assert!(
-            strategies.contains(&"admin_managed_channels"),
-            "operator token should see Slack admin channel setup: {body}"
-        );
-        assert!(
-            strategies.contains(&"oauth"),
-            "operator token should see personal Slack OAuth connection: {body}"
-        );
-
-        runtime.shutdown().await.expect("runtime shuts down");
-    }
-
-    #[tokio::test]
-    async fn slack_connectable_channels_hide_admin_action_from_sso_session_token() {
-        let (runtime, _root) = runtime().await;
-        let mounts = build_slack_host_beta_mounts(&runtime, config_without_channel_routes())
-            .expect("mounts");
-        let bundle = build_webui_services_with_slack_host_beta_mounts(
-            &runtime,
-            None,
-            Some(&mounts),
-            SlackOperatorRouteVisibility::Visible,
-        )
-        .expect("webui bundle");
-        let app = webui_v2_app(
-            bundle,
-            WebuiServeConfig::new(
-                TenantId::new(TENANT).expect("tenant"),
-                Arc::new(MixedSessionAndOperatorAuthenticator),
-                Vec::new(),
-            )
-            .with_default_agent_id(AgentId::new(AGENT).expect("agent"))
-            .with_default_project_id(ProjectId::new(PROJECT).expect("project"))
-            .with_slack_channel_routes(mounts.channel_routes),
-        )
-        .expect("webui app");
-
-        let response = app
-            .clone()
-            .oneshot(
-                Request::builder()
-                    .method("GET")
-                    .uri("/api/webchat/v2/channels/connectable")
-                    .header("authorization", "Bearer session-token")
-                    .body(Body::empty())
-                    .expect("request builds"),
-            )
-            .await
-            .expect("route responds");
-
-        assert_eq!(response.status(), StatusCode::OK);
-        let body = axum::body::to_bytes(response.into_body(), 64 * 1024)
-            .await
-            .expect("response body");
-        let body: serde_json::Value = serde_json::from_slice(&body).expect("response json");
-        let session_strategies: Vec<_> = body["channels"]
-            .as_array()
-            .expect("channels")
-            .iter()
-            .map(|channel| channel["strategy"].as_str().expect("strategy"))
-            .collect();
-        assert!(
-            !session_strategies.contains(&"admin_managed_channels"),
-            "SSO session token should not see Slack admin setup: {body}"
-        );
-        assert!(
-            session_strategies.contains(&"oauth"),
-            "SSO session token should see personal Slack OAuth connection: {body}"
-        );
-
-        let response = app
-            .oneshot(
-                Request::builder()
-                    .method("GET")
-                    .uri("/api/webchat/v2/channels/connectable")
-                    .header("authorization", "Bearer operator-token")
-                    .body(Body::empty())
-                    .expect("request builds"),
-            )
-            .await
-            .expect("route responds");
-
-        assert_eq!(response.status(), StatusCode::OK);
-        let body = axum::body::to_bytes(response.into_body(), 64 * 1024)
-            .await
-            .expect("response body");
-        let body: serde_json::Value = serde_json::from_slice(&body).expect("response json");
-        let operator_strategies: Vec<_> = body["channels"]
-            .as_array()
-            .expect("channels")
-            .iter()
-            .map(|channel| channel["strategy"].as_str().expect("strategy"))
-            .collect();
-        assert!(
-            operator_strategies.contains(&"admin_managed_channels"),
-            "operator token should see Slack admin setup: {body}"
-        );
-
-        runtime.shutdown().await.expect("runtime shuts down");
-    }
-
-    #[tokio::test]
-    async fn slack_channel_route_admin_is_reachable_through_webui_v2_app() {
-        let (runtime, _root) = runtime().await;
-        let mounts = build_slack_host_beta_mounts(&runtime, config_without_channel_routes())
-            .expect("mounts");
-        let bundle = build_webui_services_with_slack_host_beta_mounts(
-            &runtime,
-            None,
-            Some(&mounts),
-            SlackOperatorRouteVisibility::Visible,
-        )
-        .expect("webui bundle");
-        let app = webui_v2_app(
-            bundle,
-            WebuiServeConfig::new(
-                TenantId::new(TENANT).expect("tenant"),
-                Arc::new(OperatorTokenAuthenticator),
-                Vec::new(),
-            )
-            .with_default_agent_id(AgentId::new(AGENT).expect("agent"))
-            .with_default_project_id(ProjectId::new(PROJECT).expect("project"))
-            .with_slack_channel_routes(mounts.channel_routes),
-        )
-        .expect("webui app");
-
-        let unauthenticated = app
-            .clone()
-            .oneshot(
-                Request::builder()
-                    .method("PUT")
-                    .uri(WEBUI_V2_CHANNELS_SLACK_ROUTES_PATH)
-                    .header("content-type", "application/json")
-                    .body(Body::from(format!(
-                        r#"{{"channel_id":"C0HOST","subject_user_id":"{SHARED_SUBJECT}"}}"#
-                    )))
-                    .expect("unauthenticated request builds"),
-            )
-            .await
-            .expect("unauthenticated route responds");
-        assert_eq!(unauthenticated.status(), StatusCode::UNAUTHORIZED);
-
-        let empty_list = app
-            .clone()
-            .oneshot(
-                Request::builder()
-                    .method("GET")
-                    .uri(WEBUI_V2_CHANNELS_SLACK_ROUTES_PATH)
-                    .header("authorization", "Bearer operator-token")
-                    .body(Body::empty())
-                    .expect("empty list request builds"),
-            )
-            .await
-            .expect("empty list route responds");
-        assert_eq!(empty_list.status(), StatusCode::OK);
-        let body = axum::body::to_bytes(empty_list.into_body(), 64 * 1024)
-            .await
-            .expect("empty list body");
-        let body: serde_json::Value = serde_json::from_slice(&body).expect("empty list json");
-        assert_eq!(body["routes"], serde_json::json!([]));
-        assert_eq!(body["next_cursor"], serde_json::Value::Null);
-
-        let upsert = app
-            .clone()
-            .oneshot(
-                Request::builder()
-                    .method("PUT")
-                    .uri(WEBUI_V2_CHANNELS_SLACK_ROUTES_PATH)
-                    .header("authorization", "Bearer operator-token")
-                    .header("content-type", "application/json")
-                    .body(Body::from(format!(
-                        r#"{{"channel_id":"C0HOST","subject_user_id":"{SHARED_SUBJECT}"}}"#
-                    )))
-                    .expect("upsert request builds"),
-            )
-            .await
-            .expect("upsert route responds");
-        assert_eq!(upsert.status(), StatusCode::OK);
-
-        let list = app
-            .clone()
-            .oneshot(
-                Request::builder()
-                    .method("GET")
-                    .uri(WEBUI_V2_CHANNELS_SLACK_ROUTES_PATH)
-                    .header("authorization", "Bearer operator-token")
-                    .body(Body::empty())
-                    .expect("list request builds"),
-            )
-            .await
-            .expect("list route responds");
-        assert_eq!(list.status(), StatusCode::OK);
-        let body = axum::body::to_bytes(list.into_body(), 64 * 1024)
-            .await
-            .expect("list body");
-        let body: serde_json::Value = serde_json::from_slice(&body).expect("list json");
-        assert_eq!(body["routes"][0]["channel_id"], "C0HOST");
-        assert_eq!(body["routes"][0]["subject_user_id"], SHARED_SUBJECT);
-
-        let delete = app
-            .clone()
-            .oneshot(
-                Request::builder()
-                    .method("DELETE")
-                    .uri(WEBUI_V2_CHANNELS_SLACK_ROUTES_PATH)
-                    .header("authorization", "Bearer operator-token")
-                    .header("content-type", "application/json")
-                    .body(Body::from(r#"{"channel_id":"C0HOST"}"#))
-                    .expect("delete request builds"),
-            )
-            .await
-            .expect("delete route responds");
-        assert_eq!(delete.status(), StatusCode::OK);
-        let body = axum::body::to_bytes(delete.into_body(), 64 * 1024)
-            .await
-            .expect("delete body");
-        let body: serde_json::Value = serde_json::from_slice(&body).expect("delete json");
-        assert_eq!(body["deleted"], true);
-
-        let list_after_delete = app
-            .oneshot(
-                Request::builder()
-                    .method("GET")
-                    .uri(WEBUI_V2_CHANNELS_SLACK_ROUTES_PATH)
-                    .header("authorization", "Bearer operator-token")
-                    .body(Body::empty())
-                    .expect("list request builds"),
-            )
-            .await
-            .expect("list route responds");
-        assert_eq!(list_after_delete.status(), StatusCode::OK);
-        let body = axum::body::to_bytes(list_after_delete.into_body(), 64 * 1024)
-            .await
-            .expect("list body");
-        let body: serde_json::Value = serde_json::from_slice(&body).expect("list json");
-        assert_eq!(body["routes"], serde_json::json!([]));
-
-        runtime.shutdown().await.expect("runtime shuts down");
-    }
-
-    #[tokio::test]
-    async fn slack_channel_routes_mount_for_sso_operator_authenticator() {
-        let (runtime, _root) = runtime().await;
-        let mounts = build_slack_host_beta_mounts(&runtime, config_without_channel_routes())
-            .expect("mounts");
-        let bundle = build_webui_services_with_slack_host_beta_mounts(
-            &runtime,
-            None,
-            Some(&mounts),
-            SlackOperatorRouteVisibility::Hidden,
-        )
-        .expect("webui bundle");
-        let caller = WebUiAuthenticatedCaller::new(
-            TenantId::new(TENANT).expect("tenant"),
-            UserId::new(USER).expect("user"),
-            Some(AgentId::new(AGENT).expect("agent")),
-            Some(ProjectId::new(PROJECT).expect("project")),
-        );
-        let connectable = bundle
-            .api
-            .list_connectable_channels(caller)
-            .await
-            .expect("connectable channels");
-        assert!(
-            connectable
-                .channels
-                .iter()
-                .any(|channel| channel.strategy == RebornChannelConnectStrategy::OAuth),
-            "non-operator WebUI should still advertise personal Slack OAuth"
-        );
-        assert!(
-            connectable
-                .channels
-                .iter()
-                .all(|channel| channel.strategy
-                    != RebornChannelConnectStrategy::AdminManagedChannels),
-            "non-operator WebUI must not advertise Slack admin channel management"
-        );
-        let app = webui_v2_app(
-            bundle,
-            WebuiServeConfig::new(
-                TenantId::new(TENANT).expect("tenant"),
-                Arc::new(OperatorTokenAuthenticator),
-                Vec::new(),
-            )
-            .with_default_agent_id(AgentId::new(AGENT).expect("agent"))
-            .with_default_project_id(ProjectId::new(PROJECT).expect("project"))
-            .with_slack_channel_routes(mounts.channel_routes),
-        )
-        .expect("webui app");
-
-        let response = app
-            .oneshot(
-                Request::builder()
-                    .method("GET")
-                    .uri(WEBUI_V2_CHANNELS_SLACK_ROUTES_PATH)
-                    .header("authorization", "Bearer operator-token")
-                    .body(Body::empty())
-                    .expect("request builds"),
-            )
-            .await
-            .expect("route responds");
-        assert_eq!(response.status(), StatusCode::OK);
-
-        runtime.shutdown().await.expect("runtime shuts down");
-    }
-
-    #[tokio::test]
     async fn slack_tools_extension_lists_as_auth_required_until_personal_oauth_connected() {
         let (runtime, _root) = runtime().await;
         let mounts =
@@ -2664,48 +2166,6 @@ mod tests {
                 .all(|extension| extension.package_ref.id.as_str() != "slack_bot"),
             "the operator-provisioned Slack bot channel must stay hidden from the user extension catalog"
         );
-
-        runtime.shutdown().await.expect("runtime shuts down");
-    }
-
-    #[tokio::test]
-    async fn slack_channel_routes_not_mounted_when_operator_route_visibility_is_hidden() {
-        let (runtime, _root) = runtime().await;
-        let mounts = build_slack_host_beta_mounts(&runtime, config_without_channel_routes())
-            .expect("mounts");
-        let bundle = build_webui_services_with_slack_host_beta_mounts(
-            &runtime,
-            None,
-            Some(&mounts),
-            SlackOperatorRouteVisibility::Hidden,
-        )
-        .expect("webui bundle");
-        let app = webui_v2_app(
-            bundle,
-            WebuiServeConfig::new(
-                TenantId::new(TENANT).expect("tenant"),
-                Arc::new(HiddenOperatorRouteAuthenticator),
-                Vec::new(),
-            )
-            .with_default_agent_id(AgentId::new(AGENT).expect("agent"))
-            .with_default_project_id(ProjectId::new(PROJECT).expect("project"))
-            .with_slack_channel_routes(mounts.channel_routes),
-        )
-        .expect("webui app");
-
-        let response = app
-            .oneshot(
-                Request::builder()
-                    .method("GET")
-                    .uri(WEBUI_V2_CHANNELS_SLACK_ROUTES_PATH)
-                    .header("authorization", "Bearer operator-token")
-                    .body(Body::empty())
-                    .expect("request builds"),
-            )
-            .await
-            .expect("route responds");
-
-        assert_eq!(response.status(), StatusCode::NOT_FOUND);
 
         runtime.shutdown().await.expect("runtime shuts down");
     }
@@ -3775,77 +3235,6 @@ mod tests {
         assert_eq!(
             operator_targets.targets[0].target.display_name.as_str(),
             "Slack channel C0HOST"
-        );
-
-        runtime.shutdown().await.expect("runtime shuts down");
-    }
-
-    #[tokio::test]
-    async fn slack_host_beta_admin_routes_feed_outbound_target_provider() {
-        let (runtime, _root) = runtime().await;
-        let mounts = build_slack_host_beta_mounts(&runtime, config_without_channel_routes())
-            .expect("mounts");
-        let bundle = build_webui_services_with_slack_host_beta_mounts(
-            &runtime,
-            None,
-            Some(&mounts),
-            SlackOperatorRouteVisibility::Visible,
-        )
-        .expect("webui bundle");
-        let app = webui_v2_app(
-            bundle.clone(),
-            WebuiServeConfig::new(
-                TenantId::new(TENANT).expect("tenant"),
-                Arc::new(OperatorTokenAuthenticator),
-                Vec::new(),
-            )
-            .with_default_agent_id(AgentId::new(AGENT).expect("agent"))
-            .with_default_project_id(ProjectId::new(PROJECT).expect("project"))
-            .with_slack_channel_routes(mounts.channel_routes),
-        )
-        .expect("webui app");
-
-        let save = app
-            .oneshot(
-                Request::builder()
-                    .method("PUT")
-                    .uri(WEBUI_V2_CHANNELS_SLACK_ALLOWED_PATH)
-                    .header("authorization", "Bearer operator-token")
-                    .header("content-type", "application/json")
-                    .body(Body::from(r#"{"channel_ids":["C0DYNAMIC"]}"#))
-                    .expect("save request builds"),
-            )
-            .await
-            .expect("save route responds");
-        assert_eq!(save.status(), StatusCode::OK);
-        let body = axum::body::to_bytes(save.into_body(), 64 * 1024)
-            .await
-            .expect("save body");
-        let body: serde_json::Value = serde_json::from_slice(&body).expect("save json");
-        let subject_user_id = body["channels"][0]["subject_user_id"]
-            .as_str()
-            .expect("assigned subject");
-        let caller = WebUiAuthenticatedCaller::new(
-            TenantId::new(TENANT).expect("tenant"),
-            UserId::new(subject_user_id).expect("subject user"),
-            Some(AgentId::new(AGENT).expect("agent")),
-            Some(ProjectId::new(PROJECT).expect("project")),
-        );
-
-        let targets = bundle
-            .api
-            .list_outbound_delivery_targets(caller)
-            .await
-            .expect("dynamic route target list");
-
-        assert_eq!(targets.targets.len(), 1);
-        assert_eq!(
-            targets.targets[0].target.target_id.as_str(),
-            "slack:shared-channel:T0HOST:C0DYNAMIC"
-        );
-        assert_eq!(
-            targets.targets[0].target.display_name.as_str(),
-            "Slack channel C0DYNAMIC"
         );
 
         runtime.shutdown().await.expect("runtime shuts down");
