+++ crates/ironclaw_reborn_webui_ingress/Cargo.toml
@@ -22,6 +22,21 @@ default = []
 # and the O(n) constant-time lookup would scale poorly under real
 # user load. Local dev and tests opt in by enabling this feature.
 dev-in-memory-session = []
+# Forward the Slack host-beta surface. `webui_v2_app` (folded up here from
+# composition) mounts the Slack personal-OAuth setup + channel-route admin
+# surface under this feature; it forwards to composition, which owns the Slack
+# adapter and the route/mount builders this crate consumes through its public
+# surface.
+slack-v2-host-beta = ["ironclaw_reborn_composition/slack-v2-host-beta"]
+# Forward the OpenAI-compat host-beta surface: the bearer-auth middleware stamps
+# an `OpenAiCompatActorScope` onto each verified caller for protected
+# OpenAI-compatible route mounts. Pulls in the openai-compat scope/caller types
+# and the product-adapters bearer-evidence helper the middleware calls.
+openai-compat-beta = [
+    "dep:ironclaw_reborn_openai_compat",
+    "ironclaw_reborn_composition/openai-compat-beta",
+    "ironclaw_reborn_openai_compat/openai-compat-beta",
+]
 
 [dependencies]
 async-stream = "0.3"
@@ -37,6 +52,13 @@ ironclaw_host_api = { path = "../ironclaw_host_api" }
 # only the `RebornServicesApi` facade + ingress/error vocabulary.
 ironclaw_product_workflow = { path = "../ironclaw_product_workflow" }
 ironclaw_reborn_composition = { path = "../ironclaw_reborn_composition", features = ["webui-v2-beta"] }
+# Optional, enabled by `openai-compat-beta`: the moved `webui_v2_app` auth
+# middleware stamps an `OpenAiCompatActorScope` onto verified callers for
+# protected OpenAI-compatible route mounts. The tenant-scoped bearer evidence it
+# pairs with the scope comes from `ironclaw_product_adapters` via composition's
+# facade (`ironclaw_reborn_composition::mark_bearer_token_verified_for_tenant`) —
+# ingress must not depend on product_adapters directly (architecture boundary).
+ironclaw_reborn_openai_compat = { path = "../ironclaw_reborn_openai_compat", optional = true }
 # Sliding-window per-route rate limiter (moved from composition's webui_rate_limit).
 lru = "0.18"
 # axum/tower-http middleware stack used by `webui_serve` (moved from composition).
@@ -77,6 +99,13 @@ ironclaw_product_workflow = { path = "../ironclaw_product_workflow", features = 
 # Handler contract test (moved from ironclaw_webui_v2) constructs
 # `ProductOutboundEnvelope` fixtures with the canonical newtype constructors.
 ironclaw_product_adapters = { path = "../ironclaw_product_adapters" }
+# `tests/slack_host_beta_webui_v2.rs` (gated on `slack-v2-host-beta`) drives the
+# Slack host-beta surface composed into `webui_v2_app`. It builds a local-dev
+# runtime through composition's public builders, so it needs composition's
+# `test-support` + `libsql` seams plus the model-gateway / capability-port traits.
+ironclaw_reborn_composition = { path = "../ironclaw_reborn_composition", features = ["test-support", "libsql"] }
+ironclaw_loop_host = { path = "../ironclaw_loop_host" }
+tempfile = "3"
 ironclaw_threads = { path = "../ironclaw_threads" }
 ironclaw_turns = { path = "../ironclaw_turns" }
 # Generate test RSA keypairs at test time so the OIDC e2e test can
@@ -91,16 +120,3 @@ tokio = { version = "1", features = ["macros", "rt", "rt-multi-thread", "net", "
 # Real WebSocket client for the moved `webui_v2_handlers_contract` stream_events_ws
 # frame-content test (oneshot cannot complete a true WS upgrade).
 tokio-tungstenite = { version = "0.29", default-features = false, features = ["connect"] }
-
-# The WebChat v2 gateway assembly (`webui_serve.rs`, folded up from composition)
-# still carries `#[cfg(feature = "slack-v2-host-beta")]` /
-# `#[cfg(feature = "openai-compat-beta")]` blocks that reference
-# composition-internal Slack / OpenAI-compat surfaces. Those features are NOT
-# defined on this crate (wiring the Slack/OpenAI host surface through ingress is
-# deferred — see TODO(slack-merge) in webui_serve.rs), so the blocks compile out
-# unconditionally. Declare the cfg values as known so the build stays
-# warning-clean under `-D warnings` without silently enabling the dead paths.
-[lints.rust]
-unexpected_cfgs = { level = "warn", check-cfg = [
-    'cfg(feature, values("slack-v2-host-beta", "openai-compat-beta"))',
-] }
